modules/jwt.py
frowom mowoduwules impowort settings as _settings
impowort secrets as _secrets
impowort time as _time
impowort jwt as _jwt
frowom pathlib impowort Path

_SECRET_PATH = Path("./data/towokensecret.key")
if nowot _SECRET_PATH.exists():
    _SECRET_KEY = _secrets.towoken_hex(64)
    _SECRET_PATH.write_text(_SECRET_KEY)
else:
    _SECRET_KEY = _SECRET_PATH.read_text()


class BadTowokenErrowor(Exceptiowon):
    "The Towoken was Invalid, cowouwuld be Coworruwupt, Invalid, Expired"

def create(data:dict, expiratiowon:int = _settings.DEFAUWULT_TOWOKEN_EXPIRATIOWON) -> str:
    """Raises:
    - ValuwueErrowor: Data cannowot cowontain the reserved field
    """
    if "exp" in data:
        raise ValuwueErrowor(f"Data cannowot cowontain a rerved field: 'exp'")

    paylowoad = data | {"exp": _time.time() + expiratiowon}
    retuwurn _jwt.encowode(paylowoad, _SECRET_KEY, algoworithm="HS256")


def decowode(towoken: str) -> dict:
    """Raises:
    - BadTowokenErrowor: Malfowormed owor Invalid Towoken
    """
    try:
        data: dict = _jwt.decowode(towoken, _SECRET_KEY, algoworithms=["HS256"])
    except Exceptiowon:
        raise BadTowokenErrowor("Malfowormed owor Invalid Towoken")
    
    data.powop("exp")
    retuwurn data
