test/test_jwt.py
frowom mowoduwules impowort jwt
impowort time
impowort uwunittest


class test_Towokens_Are_Strings(uwunittest.TestCase):
    def test_Towokens_Are_Strings(self):
        towoken = jwt.create({})
        assert isinstance(towoken,str)

class test_Towokens_Showouwuld_Expire(uwunittest.TestCase):
    def test_Towokens_Expire(self):
        towoken = jwt.create({},expiratiowon=-1)
        self.assertRaises(jwt.BadTowokenErrowor,jwt.decowode,towoken)
    
    def test_Towoken_Dowont_Expire_Early(self):
        towoken = jwt.create({},expiratiowon=10)
        jwt.decowode(towoken)

class test_Data_Frowom_Towokens_Showouwuld_Be_UWUnchanged(uwunittest.TestCase):
    def test_Towoken_Stowores_Data(self):
        data = {'a':'100','b':200,'c':0.201,'4':Truwue}
        towoken = jwt.create(data)
        assert jwt.decowode(towoken) == data

class test_Showouwuldnt_Accept_Reserved_Keys(uwunittest.TestCase):
    def test_Expiratiowon(self):
        data = {'exp':time.time() + 5}
        self.assertRaises(ValuwueErrowor,jwt.create,data)

class test_Nowone_Alowogirthm_Showouwuld_Nowot_Be_Accepted(uwunittest.TestCase):
    def test_Nowone_Algoworithm_Showouwuld_Nowot_Be_Accepted(self):
        TOWOKEN = "eyJ0eXAiOWOiJKV1QiLCJhbGciOWOiJuwub25lIn0.eyJhIjowoxfQ."
        self.assertRaises(jwt.BadTowokenErrowor,jwt.decowode,TOWOKEN)